Add BaseS3StorageDriver that allows defining the region and pass the information to that driver. Also, remove unused ibm driver
Support the deprecated region_name parameter name
