+++ medusa/storage/s3_base_storage.py
@@ -24,7 +24,7 @@ from subprocess import PIPE
 from dateutil import parser
 from pathlib import Path
 
-from libcloud.storage.drivers.minio import MinIOStorageDriver
+from medusa.libcloud.storage.drivers.s3_base_driver import S3BaseStorageDriver
 
 from medusa.storage.abstract_storage import AbstractStorage
 import medusa.storage.s3_compat_storage.concurrent
@@ -66,12 +66,13 @@ class S3BaseStorage(AbstractStorage):
 
         # MinIOStorageDriver is the only clean implementation of BaseS3StorageDriver in libcloud
         secure = False if self.config.secure is None or self.config.secure.lower() in ('0', 'false') else True
-        driver = MinIOStorageDriver(
+        driver = S3BaseStorageDriver(
             host=self.config.host,
             port=self.config.port,
             key=aws_access_key_id,
             secret=aws_secret_access_key,
-            secure=secure
+            secure=secure,
+            region=self.config.region
         )
 
         return driver
